+++ Source/EventFlow/ReadStores/InMemory/IInMemoryReadStore.cs
@@ -20,10 +20,18 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-namespace EventFlow.ReadStores.MsSql
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace EventFlow.ReadStores.InMemory
 {
-    public interface IMssqlReadModelStore<TReadModel> : IReadModelStore<TReadModel>
-        where TReadModel : class, IMssqlReadModel, new()
+    public interface IInMemoryReadStore<TReadModel> : IReadModelStore<TReadModel>
+        where TReadModel : class, IReadModel, new()
     {
+        Task<IReadOnlyCollection<TReadModel>> FindAsync(
+            Predicate<TReadModel> predicate,
+            CancellationToken cancellationToken);
     }
 }