+++ Source/EventFlow/ReadStores/InMemory/Queries/InMemoryQuery.cs
@@ -26,7 +26,7 @@ using EventFlow.Queries;
 
 namespace EventFlow.ReadStores.InMemory.Queries
 {
-    public class InMemoryQuery<TReadModel> : IQuery<IEnumerable<TReadModel>>
+    public class InMemoryQuery<TReadModel> : IQuery<IReadOnlyCollection<TReadModel>>
         where TReadModel : IReadModel, new()
     {
         public Predicate<TReadModel> Query { get; private set; }